% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_stats_functions.R
\name{addAverages}
\alias{addAverages}
\title{Add mean and median to plot}
\usage{
addAverages(
  plot,
  df_stats,
  ymin,
  averages_point_size = 3,
  mean_shape = 23,
  mean_fill = "white",
  mean_colour = "black",
  mean_stroke = 0.8,
  median_shape = 21,
  median_fill = "black",
  median_colour = "black",
  show_mean = T,
  show_median = T,
  averages_opacity = 0.8
)
}
\arguments{
\item{ymin}{position of the mean symbol on the y-axis, which equals the lower y limit of the rectangle plotted for mean +- SD}

\item{averages_point_size}{point size for the mean, median will be 20 percent smaller, defaults to 3}

\item{mean_shape}{point shape for the mean, defaults to 23 (a diamond)}

\item{mean_fill}{the fill colour for the mean, defaults to 'white'}

\item{mean_colour}{outline colour for the mean, defaults to 'black'}

\item{mean_stroke}{size of the outer line of the symbol for the mean, defaults to 0.8}

\item{median_shape}{point shape for the median, defaults to 21 (a circle)}

\item{median_fill}{the fill colour for the median, defaults to 'black'}

\item{median_colour}{outline colour for the median, defaults to 'black'}

\item{show_mean}{logical, false if the mean is not to be added to the plot, defaults to TRUE}

\item{show_median}{logical, false if the median is not to be added to the plot, defaults to TRUE}

\item{averages_opacity}{alpha value for the mean and median points, numeric between 0 and 1, defaults to 0.8}
}
\value{
a ggplot plot with a point for each groups mean and median added (depending on parameters)
}
\description{
Adds points to histogram, ridge plot or sea stack plot for mean and median. Defaults to
diamond for mean, and a circle for the median, of symbol size slightly smaller than the
symbol size for the mean. Positions of the mean at the lower limit of the SD bar (so that
the SD bar covers one half of the diamond) and median on the x-axis (y = 0). Symbol size
for mean defaults to 3, while the median is 20 percent smaller.
}
