% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_plots_functions.R
\name{plotFormatting}
\alias{plotFormatting}
\title{Improve the formatting of the basic skyline plot}
\usage{
plotFormatting(
  plot,
  legend = F,
  panel_spacing = 0.5,
  x_title_size = 12,
  y_title_size = 12,
  x_title_face = "bold",
  y_title_face = "bold",
  axis_text_size = 12,
  group_label_size = 12,
  axis_text_face = NULL,
  strip_text_face = NULL,
  y_lab = "Count",
  brewer_fill = NULL,
  brewer_colour = NULL
)
}
\arguments{
\item{plot}{basic plot from \code{\link{skylinePlot}}}

\item{legend}{logical, true to include legend and false to exclude, default value = F.}

\item{panel_spacing}{value for the spacing between faceted panels in cm, default value = 0.5cm}

\item{x_title_size}{text size for x-axis label, default value = 12}

\item{y_title_size}{text size for y-axis label, default value = 12}

\item{x_title_face}{text face for x-axis label, default value = 'bold'}

\item{y_title_face}{text face for y-axis label, default value = 'bold'}

\item{axis_text_size}{text size for axis text, default value = 12}

\item{group_label_size}{text face for group labels, default value = NULL}

\item{axis_text_face}{text face for axis text, default value = NULL}

\item{y_lab}{y-axis label, defaults to 'Count'}

\item{brewer_fill}{if you wish to have a brewer colour palette, as histogram fill, set the palette name here, if not leave as NULL (default)}

\item{brewer_colour}{if you wish to have a brewer colour palette, as histogram colour, set the palette name here, if not leave as NULL (default)}
}
\description{
The default theme is not really cool, so will will make it all a bit nicer
potential to choose nicer colors using brewer_fill and brewer_colour variables
}
