% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_plots_functions.R
\name{plotStats}
\alias{plotStats}
\title{Add summary statistics to a plot}
\usage{
plotStats(
  plot,
  SD_fill = "grey30",
  SD_colour = NA,
  SD_size = NULL,
  CI_colour = "red",
  CI_size = 2,
  CI_width = 1,
  show_CI = T,
  show_SD = T,
  averages_point_size = 3.5,
  mean_shape = 23,
  mean_fill = "white",
  mean_colour = "black",
  median_shape = 21,
  median_fill = "black",
  median_colour = "black",
  show_mean = T,
  show_median = T,
  averages_opacity = 0.8
)
}
\description{
Add mean, median, standard deviation and/or confidence intervals to a histogram,
skyline plot, density plot, boxplot, violin plot or sea stack plot.
NOTE: histogram must be first layer of \code{plot} for this to work unless summary
statistics from \code{\link{summaryStats}} are added
}
