% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_ridge_plot_functions.R
\name{seaStackPlotRidge}
\alias{seaStackPlotRidge}
\title{Sea Stack plot with clean inside}
\usage{
seaStackPlotRidge(
  df,
  value,
  group,
  colour = "grey50",
  fill = "grey90",
  alpha = 0.5,
  bins = NULL,
  binwidth = NULL,
  confidence_interval = 0.95,
  legend = F,
  panel_spacing = 0.5,
  x_title_size = 12,
  y_title_size = 12,
  x_title_face = "bold",
  y_title_face = "bold",
  axis_text_size = 12,
  group_label_size = 12,
  axis_text_face = NULL,
  group_label_face = NULL,
  y_lab = "Count",
  SD_fill = "grey30",
  SD_colour = NA,
  SD_size = NULL,
  CI_colour = "red",
  CI_size = 2,
  CI_width = 1,
  show_CI = T,
  show_SD = T,
  averages_point_size = 3.5,
  mean_shape = 23,
  mean_fill = "white",
  mean_colour = "black",
  median_shape = 21,
  median_fill = "black",
  median_colour = "black",
  show_mean = T,
  show_median = T,
  averages_opacity = 0.8,
  df_stats = NULL,
  vertical = T,
  mirrored = T
)
}
\arguments{
\item{df}{data frame or tibble of values with at least column for grouping variable and column for values}

\item{value}{name of the column containing values, should be given as a character string}

\item{group}{name of the column containing grouping variable, should be given as a character string}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{confidence_interval}{desired size of the confidence interval, value between 0 and 1 e.g. 0.95 for 95% (default)}

\item{legend}{logical, true to include legend and false to exclude, default value = F.}

\item{panel_spacing}{value for the spacing between faceted panels in cm, default value = 0.5cm}

\item{x_title_size}{text size for x-axis label, default value = 12}

\item{y_title_size}{text size for y-axis label, default value = 12}

\item{x_title_face}{text face for x-axis label, default value = 'bold'}

\item{y_title_face}{text face for y-axis label, default value = 'bold'}

\item{axis_text_size}{text size for axis text, default value = 12}

\item{group_label_size}{text face for group labels, default value = NULL}

\item{axis_text_face}{text face for axis text, default value = NULL}

\item{y_lab}{y-axis label, defaults to 'Count'}

\item{SD_fill}{character string specifying the fill for the standard deviation rectangle}

\item{SD_colour}{character string specifying the colour for the standard deviation rectangle}

\item{SD_size}{height of standard deviation rectangle, gets set to a tenth of the
height of the tallest bin unless specified. NOTE: this does not work on data that
has been processed before being plotted}

\item{CI_colour}{character string specifying the colour for the confidence interval lines}

\item{show_CI}{logical, if false will not plot confidence interval marks, defaults to TRUE}

\item{show_SD}{logical, if false will not plot standard deviation rectangle, defaults to TRUE}

\item{averages_point_size}{point size for the mean, median will be 20 percent smaller, defaults to 3.5}

\item{mean_shape}{point shape for the mean, defaults to 23 (a diamond)}

\item{mean_fill}{the fill colour for the mean, defaults to 'white'}

\item{mean_colour}{outline colour for the mean, defaults to 'black'}

\item{median_shape}{point shape for the median, defaults to 21 (a circle)}

\item{median_fill}{the fill colour for the median, defaults to 'black'}

\item{median_colour}{outline colour for the median, defaults to 'black'}

\item{show_mean}{logical, false if the mean is not to be added to the plot, defaults to TRUE}

\item{show_median}{logical, false if the median is not to be added to the plot, defaults to TRUE}

\item{averages_opacity}{alpha value for the mean and median points, numeric between 0 and 1, defaults to 0.8}

\item{df_stats}{summary statistics for plot, if NULL (default) will calculate from data behind
plot. NOTE: this does not work on data that has been processed before being plotted}

\item{vertical}{logical, if true the plot will be flipped vertically, if false it'll remain horizontal, defaults to TRUE}

\item{mirrored}{logical, if true the plot will be mirrored so histogram is on the left hand side, else it'll stay on the right, defaults to TRUE}
}
\value{
a ggplot sea stack plot with no lines within the histogram but an outline around it
}
\description{
based off of geom_step and geom_hist
}
