% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underlying_functions.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, na.rm = F)
}
\arguments{
\item{x}{a vector of values for which the SEM is to be calculated}

\item{na.rm}{logical, TRUE to remove NA values from x (that will result in NA being returned),
defaults to FALSE}
}
\value{
standard error of the mean for x (numeric)
}
\description{
Calculates the standard error of the mean for a numeric vector, x
}
