% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_histogram_plot_functions.R
\name{skylinePlot}
\alias{skylinePlot}
\title{creates a basic horizontal skyline plot}
\usage{
skylinePlot(
  df,
  value,
  group,
  colour = NULL,
  fill = "blue",
  alpha = 0.7,
  bins = NULL,
  binwidth = NULL
)
}
\arguments{
\item{df}{data frame or tibble of values with at least column for grouping variable and column for values}

\item{value}{name of the column containing values, should be given as a character string}

\item{group}{name of the column containing grouping variable, should be given as a character string}

\item{colour}{character string vector of colour(s) for outline of histogram, will be repeated across groups if too short. Default is NULL.}

\item{fill}{character string vector of fill for histogram, will be repeated across groups if too short. Default is 'blue'.}

\item{alpha}{alpha value for histogram, between 0 and 1. Default is 0.7.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}
}
\value{
a ggplot object with faceted histograms
}
\description{
created a plot made up of a histogram for each group stacked vertically, split by \code{\link{facet_grid}}
}
\examples{
skylinePlot(df = InsectSprays,
value = 'weight',
group = 'Diet',
fill = c('blue', 'yellow', 'green'),
binwidth = 10)

}
