% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_shared_plot_functions.R
\name{verticalPlot}
\alias{verticalPlot}
\title{Flip plots to be vertical}
\usage{
verticalPlot(plot, vertical = T, mirrored = T)
}
\arguments{
\item{plot}{a histogram or ridge plot ggplot item}

\item{vertical}{logical, if true the plot will be flipped vertically, if false it'll remain horizontal, defaults to TRUE}

\item{mirrored}{logical, if true the plot will be mirrored so histogram is on the left hand side, else it'll stay on the right, defaults to TRUE}
}
\value{
a ggplot plot that has been made vertical and mirrored
}
\description{
Turn plots to be vertical and mirror axes so your sea stack plot stands up!
}
\examples{
verticalPlot(insect_sprays_stats_plot)

}
