% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_stats_functions.R
\name{addAverages}
\alias{addAverages}
\title{Add mean and median to plot}
\usage{
addAverages(
  plot,
  df_stats,
  ymin,
  ymax,
  averages_point_size = 3.5,
  mean_shape = 23,
  mean_fill = "white",
  mean_colour = "black",
  median_shape = 21,
  median_fill = "black",
  median_colour = "black",
  show_mean = T,
  show_median = T,
  averages_opacity = 0.8
)
}
\arguments{
\item{plot}{a histogram or skyline plot ggplot item}

\item{df_stats}{summary statistics for the data, if NULL (default) will calculate from data in the ggplot item}

\item{ymin}{the lower y limit of the rectangle}

\item{ymax}{the upper y limit of the rectangle}

\item{averages_point_size}{point size for the mean, median will be 20 percent smaller, defaults to 3.5}

\item{mean_shape}{point shape for the mean, defaults to 23 (a diamond)}

\item{mean_fill}{the fill colour for the mean, defaults to 'white'}

\item{mean_colour}{outline colour for the mean, defaults to 'black'}

\item{median_shape}{point shape for the median, defaults to 21 (a circle)}

\item{median_fill}{the fill colour for the median, defaults to 'black'}

\item{median_colour}{outline colour for the median, defaults to 'black'}

\item{show_mean}{logical, false if the mean is not to be added to the plot, defaults to TRUE}

\item{show_median}{logical, false if the median is not to be added to the plot, defaults to TRUE}

\item{averages_opacity}{alpha value for the mean and median points, numeric between 0 and 1, defaults to 0.8}
}
\value{
a ggplot plot with a point for each groups mean and median added (depending on parameters)
}
\description{
Adds points to histogram, skyline plot or sea stack plot for mean and median defaults to
diamond for mean, and a circle for the median, I chose to plot a circle, of symbol size
slighty smaller than the symbol size for the mean. Positions of the mean in the middle of
the SD bar and median on the x-axis. Symbol size for mean defaults to 3.5, while the
median is 20 percent smaller.
}
