% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_plots_functions.R
\name{addCI}
\alias{addCI}
\title{Add confidence intervals to histogram or skyline plot}
\usage{
addCI(
  plot,
  df_stats = NULL,
  CI_colour = "red",
  CI_max,
  CI_min,
  CI_width,
  confidence_interval = 0.95
)
}
\arguments{
\item{plot}{a histogram or skyline plot ggplot item}

\item{df_stats}{summary statistics for the data, if NULL (default) will calculate from data in the ggplot item}

\item{CI_colour}{character string specifying the colour for the confidence interval lines}

\item{CI_max}{the top of the}

\item{CI_min}{the upper y limit of the rectangle}
}
\description{
Adds lines perpendicular to the x-axis at the
}
