% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_stats_functions.R
\name{addSD}
\alias{addSD}
\title{Add standard deviation to histogram or skyline plot}
\usage{
addSD(plot, df_stats = NULL, ymin = 0, ymax, SD_fill, SD_colour)
}
\arguments{
\item{plot}{a histogram or skyline plot ggplot item}

\item{df_stats}{summary statistics for the data, if NULL (default) will calculate from data in the ggplot item}

\item{ymin}{the lower y limit of the rectangle}

\item{ymax}{the upper y limit of the rectangle}

\item{SD_fill}{character string specifying the fill for the standard deviation rectangle}

\item{SD_colour}{character string specifying the colour for the standard deviation rectangle}
}
\value{
a ggplot plot with rectangle showing range of mean +_ one standard deviation
}
\description{
Adds a rectangle underneath the x axis showing the values within one standard deviation of the mean
}
