% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_stats_functions.R
\name{plotStats}
\alias{plotStats}
\title{Add summary statistics to a plot}
\usage{
plotStats(
  plot,
  SD_fill = "grey30",
  SD_colour = NA,
  SD_size = NULL,
  CI_colour = "red",
  CI_size = 2,
  CI_width = 1,
  show_CI = T,
  show_SD = T,
  averages_point_size = 3,
  mean_shape = 23,
  mean_fill = "white",
  mean_colour = "black",
  mean_stroke = 0.8,
  median_shape = 21,
  median_fill = "black",
  median_colour = "black",
  show_mean = T,
  show_median = T,
  averages_opacity = 0.8,
  df_stats = NULL,
  confidence_interval = 0.95
)
}
\arguments{
\item{plot}{a histogram or ridge plot ggplot item}

\item{SD_fill}{character string specifying the fill for the standard deviation rectangle}

\item{SD_colour}{character string specifying the colour for the standard deviation rectangle}

\item{SD_size}{height of standard deviation rectangle, gets set to a tenth of the
height of the tallest bin unless specified. NOTE: this does not work on data that
has been processed before being plotted}

\item{CI_colour}{character string specifying the colour for the confidence interval lines}

\item{show_CI}{logical, if false will not plot confidence interval marks, defaults to TRUE}

\item{show_SD}{logical, if false will not plot standard deviation rectangle, defaults to TRUE}

\item{averages_point_size}{point size for the mean, median will be 20 percent smaller, defaults to 3}

\item{mean_shape}{point shape for the mean, defaults to 23 (a diamond)}

\item{mean_fill}{the fill colour for the mean, defaults to 'white'}

\item{mean_colour}{outline colour for the mean, defaults to 'black'}

\item{mean_stroke}{size of the outer line of the symbol for the mean, defaults to 0.8}

\item{median_shape}{point shape for the median, defaults to 21 (a circle)}

\item{median_fill}{the fill colour for the median, defaults to 'black'}

\item{median_colour}{outline colour for the median, defaults to 'black'}

\item{show_mean}{logical, false if the mean is not to be added to the plot, defaults to TRUE}

\item{show_median}{logical, false if the median is not to be added to the plot, defaults to TRUE}

\item{averages_opacity}{alpha value for the mean and median points, numeric between 0 and 1, defaults to 0.8}

\item{df_stats}{summary statistics for plot, if NULL (default) will calculate from data behind
plot. NOTE: this does not work on data that has been processed before being plotted}

\item{confidence_interval}{desired size of the confidence interval, value between 0 and 1 e.g. 0.95 for 95% (default)}
}
\value{
a ggplot list with original plot and chosen stats added
}
\description{
Add mean, median, standard deviation and/or confidence intervals to a histogram,
ridge plot, density plot, boxplot, violin plot or sea stack plot. NOTE: main plot
must be first layer of \code{plot} for this to work unless summary statistics from
\code{\link{summaryStats}} are added as parameter \code{df_stats}
}
\examples{
plotStats(insect_sprays_formatted_plot)

}
