% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaStackR_stats_functions.R
\name{summaryStats}
\alias{summaryStats}
\title{Calculate Summary Statistics}
\usage{
summaryStats(df, group, value, confidence_interval = 0.95)
}
\arguments{
\item{df}{data frame or tibble of values with at least column for grouping variable and column for values}

\item{group}{name of the column containing grouping variable, should be given as a character string}

\item{value}{name of the column containing values, should be given as a character string}

\item{confidence_interval}{desired size of the confidence interval, value between 0 and 1 e.g. 0.95 for 95% (default)}
}
\value{
tibble with columns group (retains name of grouping variable column), \code{Mean} (mean),
\code{SD} (standard deviation), \code{Median} (median), \code{SEM} (standard error of the mean,
calculated using \code{\link{sem}}), \code{CI} (distance of confidence interval from mean),
\code{upperBound} (mean + one standard deviation) and \code{lowerBound} (mean - one standard
deviation).
}
\description{
Calculate summary stats using the package dplyr. Here, we are starting with our data (df),
and then grouping the data by the column Group. For each group, we are calculating the mean,
standard deviation (SD), median, standard error of the mean (SEM) and the distance of the
confidence intervals from the mean
}
\examples{
summaryStats(InsectSprays, 'spray', 'count', confidence_interval = 0.8)

}
